Move : stringToNumber 함수 Utils.js로 이동
