+++ src/constants/Messages.js
@@ -1,5 +1,5 @@
 const MESSAGES = Object.freeze({
-  inputBalance : '구입금액을 입력해 주세요.\n',
+  inputBalance : '\n구입금액을 입력해 주세요.\n',
   inputWinningNumbers : '\n당첨 번호를 입력해 주세요.\n',
   inputBonusNumber : '\n보너스 번호를 입력해 주세요.\n',
